refactor(sharedCheckboxAndRadioStyles): moved to internal
