+++ src/Radio/Radio.tsx
@@ -5,7 +5,7 @@ import {FormValidationStatus} from '../utils/types/FormValidationStatus'
 import {RadioGroupContext} from '../RadioGroup/RadioGroup'
 import getGlobalFocusStyles from '../_getGlobalFocusStyles'
 import {get} from '../constants'
-import {sharedCheckboxAndRadioStyles} from '../_sharedCheckboxAndRadioStyles'
+import {sharedCheckboxAndRadioStyles} from '../internal/hooks/sharedCheckboxAndRadioStyles'
 
 export type RadioProps = {
   /**
