+++ src/internal/hooks/sharedCheckboxAndRadioStyles.ts
@@ -1,5 +1,5 @@
 import {css} from 'styled-components'
-import {get} from './constants'
+import {get} from '../../constants'
 
 export const sharedCheckboxAndRadioStyles = css`
   appearance: none;
